ocr_engine/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
pydantic-settings
python-multipart
pymupdf
Pillow
gigachat
httpx
python-dotenv
pytest
tesserocr
pdf2image
PyMuPDF
openai
